requirements.api.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic-settings==2.7.0
httpx==0.28.0
requests==2.32.3
celery==5.4.0
redis==5.2.0
pytest==8.3.4
pytest-asyncio==0.24.0
